* mpn/generic/dump.c: Add a couple of casts to keep gcc quiet.
gcc complains about a ulong printf format versus a ulonglong argument,
though that code is not reached for a long long limb.

+++ mpn/generic/dump.c
@@ -50,7 +50,7 @@ mpn_dump (mp_srcptr ptr, mp_size_t size)
 	    printf ("%lX", (unsigned long) ptr[size]);
 	}
       else
-	printf ("%lX", ptr[size]);
+	printf ("%lX", (unsigned long) ptr[size]);
 
       while (size)
 	{
@@ -63,7 +63,8 @@ mpn_dump (mp_srcptr ptr, mp_size_t size)
 		(unsigned long) ptr[size]);
 	    }
 	  else
-	    printf ("%0*lX", (int) (2 * BYTES_PER_MP_LIMB), ptr[size]);
+	    printf ("%0*lX", (int) (2 * BYTES_PER_MP_LIMB),
+                    (unsigned long) ptr[size]);
 	}
       printf ("\n");
     }
